internal/wglinux: rename from internal/wgnl
