+++ internal/wglinux/doc.go
@@ -1,6 +1,6 @@
-// Package wgnl provides internal access to Linux's WireGuard generic
+// Package wglinux provides internal access to Linux's WireGuard generic
 // netlink interface.
 //
 // This package is internal-only and not meant for end users to consume.
 // Please use package wireguardctrl (an abstraction over this package) instead.
-package wgnl
+package wglinux
